requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
youtube-transcript-api==0.6.1
openai==1.6.1
python-dotenv==1.0.0
tiktoken==0.5.2
requests==2.31.0
